+++ Reach/src/utopia/reach/component/label/image/AnimatedImageLabel.scala
@@ -1,16 +1,92 @@
 package utopia.reach.component.label.image
 
+import utopia.firmament.component.stack.ConstrainableWrapper
+import utopia.firmament.model.stack.modifier.MinOptimalSizeModifier
 import utopia.flow.view.mutable.Pointer
 import utopia.flow.view.template.eventful.Flag
 import utopia.genesis.handling.action.Actor
 import utopia.genesis.image.Image
+import utopia.paradigm.angular.DirectionalRotation
 import utopia.paradigm.animation.TimedAnimation
 import utopia.paradigm.shape.shape2d.Matrix2D
+import utopia.paradigm.shape.shape2d.vector.size.Size
+import utopia.reach.component.factory.ComponentFactoryFactory.Cff
+import utopia.reach.component.factory.FromContextComponentFactoryFactory.Ccff
 import utopia.reach.component.hierarchy.ComponentHierarchy
-import utopia.reach.component.template.{ReachComponentLike, ReachComponentWrapper}
+import utopia.reach.component.template.ReachComponentWrapper
 
 import scala.concurrent.duration.{Duration, FiniteDuration}
 
+case class AnimatedImageLabelFactory(hierarchy: ComponentHierarchy,
+                                     settings: ViewImageLabelSettings = ViewImageLabelSettings.default,
+                                     loops: Boolean = true)
+	extends ViewImageLabelSettingsWrapper[AnimatedImageLabelFactory]
+{
+	// COMPUTED ------------------------
+	
+	/**
+	  * @return A copy of this factory where the label only plays the animation once
+	  */
+	def once = copy(loops = false)
+	
+	
+	// IMPLEMENTED  --------------------
+	
+	override def self: AnimatedImageLabelFactory = this
+	
+	override def withSettings(settings: ViewImageLabelSettings): AnimatedImageLabelFactory = copy(settings = settings)
+	override def *(mod: Double): AnimatedImageLabelFactory = mapSettings { _ * mod }
+	
+	
+	// OTHER    ------------------------
+	
+	/**
+	  * Creates a new animated label
+	  * @param animation Animation to play
+	  * @param transformAnimation Animated transformation to apply (optional)
+	  * @return A new animated label
+	  */
+	def apply(animation: TimedAnimation[Image], transformAnimation: Option[TimedAnimation[Matrix2D]] = None) =
+		new AnimatedImageLabel(hierarchy, animation, transformAnimation, settings, loops)
+	/**
+	  * Creates a new animated label, based on an animated transformation
+	  * @param image Image to draw
+	  * @param transform Animated transformation to apply
+	  * @return A new animated label
+	  */
+	def transforming(image: Image, transform: TimedAnimation[Matrix2D]) =
+		apply(TimedAnimation.fixed(image, transform.duration), Some(transform))
+	
+	/**
+	  * Creates a new animated label displaying a rotating image.
+	  * @param image Image to draw
+	  * @param rotation Animated rotation to apply
+	  * @param centerOrigin Whether image origin should be set to the center.
+	  *                     Default = false = keep image origin where it is.
+	  * @return A new animated label
+	  */
+	def rotating(image: Image, rotation: TimedAnimation[DirectionalRotation], centerOrigin: Boolean = false) =
+	{
+		val label = transforming(if (centerOrigin) image.withCenterOrigin else image, rotation.map(Matrix2D.rotation))
+		// Makes it so that the label won't change in size due to the rotation
+		val imageRadius = {
+			if (centerOrigin)
+				image.origin.length
+			else
+				image.bounds.corners.iterator.map { _.length }.max
+		}
+		label.addConstraint(MinOptimalSizeModifier(Size.square(imageRadius * 2)))
+		label
+	}
+}
+
+object AnimatedImageLabel extends Cff[AnimatedImageLabelFactory] with Ccff[Any, AnimatedImageLabelFactory]
+{
+	override def apply(hierarchy: ComponentHierarchy): AnimatedImageLabelFactory = AnimatedImageLabelFactory(hierarchy)
+	
+	// Ignores the context
+	override def withContext(hierarchy: ComponentHierarchy, context: Any): AnimatedImageLabelFactory = apply(hierarchy)
+}
 /**
   * A label that displays an image + transformation -based animation
   * @author Mikko Hilpinen
@@ -19,10 +95,11 @@ import scala.concurrent.duration.{Duration, FiniteDuration}
 class AnimatedImageLabel(hierarchy: ComponentHierarchy, animation: TimedAnimation[Image],
                          transformAnimation: Option[TimedAnimation[Matrix2D]],
                          settings: ViewImageLabelSettings, looping: Boolean)
-	extends ReachComponentWrapper
+	extends ReachComponentWrapper with ConstrainableWrapper
 {
 	// ATTRIBUTES   -------------------------
 	
+	// Combines the transformation from the settings with the possible animated transformation
 	private lazy val appliedSettings = Animator.transformP match {
 		case Some(transformP) =>
 			settings.mapTransformationPointer { customP =>
@@ -33,12 +110,10 @@ class AnimatedImageLabel(hierarchy: ComponentHierarchy, animation: TimedAnimatio
 					})
 				}
 			}
-		
 		case None => settings
 	}
 	
-	override protected lazy val wrapped: ReachComponentLike = ViewImageLabel(hierarchy)
-		.withSettings(appliedSettings).apply(Animator.imageP)
+	override protected lazy val wrapped = ViewImageLabel(hierarchy).withSettings(appliedSettings).apply(Animator.imageP)
 	
 	
 	// NESTED   -----------------------------
@@ -52,6 +127,7 @@ class AnimatedImageLabel(hierarchy: ComponentHierarchy, animation: TimedAnimatio
 			case None => animation.duration
 		}
 		
+		// If this is a one-time animation, supports pointer-locking
 		private val lockableAdvanceP = if (looping) None else Some(Pointer.lockable[Duration](Duration.Zero))
 		private val advanceP = lockableAdvanceP.getOrElse { Pointer.eventful[Duration](Duration.Zero) }
 		lazy val imageP = {
@@ -69,17 +145,19 @@ class AnimatedImageLabel(hierarchy: ComponentHierarchy, animation: TimedAnimatio
 		
 		override lazy val handleCondition: Flag = {
 			if (looping)
-				linkedFlag
+				hierarchy.linkPointer
 			else
-				linkedFlag && advanceP.map { _ < maxDuration }
+				hierarchy.linkPointer && advanceP.map { _ < maxDuration }
 		}
 		
 		
 		// IMPLEMENTED  ---------------------
 		
 		override def act(duration: FiniteDuration): Unit = {
+			// Advances the animation, if possible
 			val isFinished = advanceP.mutate { advance =>
 				val newAdvance = advance + duration
+				// Checks for looping / finishing
 				if (newAdvance >= maxDuration) {
 					if (looping)
 						false -> (newAdvance - maxDuration)
@@ -89,6 +167,7 @@ class AnimatedImageLabel(hierarchy: ComponentHierarchy, animation: TimedAnimatio
 				else
 					false -> newAdvance
 			}
+			// Once/if finished, locks the advance pointer
 			if (isFinished)
 				lockableAdvanceP.foreach { _.lock() }
 		}